refactor: made packages lowercase and singular
