+++ Dijstra-Algorithm-Project-INF21/src/model/Graph.java
@@ -1,4 +1,4 @@
-package Model;
+package model;
 
 import java.io.Serializable;
 import java.util.ArrayList;
