+++ Dijstra-Algorithm-Project-INF21/src/model/Way.java
@@ -1,7 +1,5 @@
-package Model;
+package model;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 /**
