+++ Dijstra-Algorithm-Project-INF21/src/service/XmlElements/Node.java
@@ -1,4 +1,4 @@
-package Services.XmlElements;
+package service.XmlElements;
 
 public class Node {
     private int id;
