+++ Dijstra-Algorithm-Project-INF21/src/service/XmlElements/Way.java
@@ -1,4 +1,4 @@
-package Services.XmlElements;
+package service.XmlElements;
 
 import java.util.ArrayList;
