--- Dijstra-Algorithm-Project-INF21/src/ui/ViewModel.java
@@ -1,20 +1,0 @@
-package ui;
-
-import java.util.ArrayList;
-
-public class ViewModel {
-
-    private Object map;
-    private Object route;
-
-    private ArrayList<Object> listeners;
-
-    public void setMap(Object map) {
-
-    }
-
-    public void setRoute(Object route) {
-
-    }
-
-}
